In light of #8255, use BIGINTs for stream_orderings

+++ synapse/storage/databases/main/schema/delta/58/17_catchup_last_successful.sql
@@ -18,4 +18,4 @@
 -- A value of NULL means that we have not sent an event successfully yet
 -- (at least, not since the introduction of this column).
 ALTER TABLE destinations
-    ADD COLUMN last_successful_stream_ordering INTEGER;
+    ADD COLUMN last_successful_stream_ordering BIGINT;
